refactor(prompts): extract implementation landing contract
